Expose carwriter and filestore by moving these out of internal
This allows these packages to be used outside storetheindex, since these may be useful to others such as index-provider.
