+++ carwriter/carwriter_test.go
@@ -15,9 +15,9 @@ import (
 	"github.com/ipld/go-ipld-prime"
 	cidlink "github.com/ipld/go-ipld-prime/linking/cid"
 	"github.com/ipni/storetheindex/api/v0/ingest/schema"
+	"github.com/ipni/storetheindex/carwriter"
 	"github.com/ipni/storetheindex/config"
-	"github.com/ipni/storetheindex/internal/carwriter"
-	"github.com/ipni/storetheindex/internal/filestore"
+	"github.com/ipni/storetheindex/filestore"
 	"github.com/ipni/storetheindex/test/util"
 	crypto "github.com/libp2p/go-libp2p/core/crypto"
 	"github.com/libp2p/go-libp2p/core/peer"
