+++ filestore/error.go
