+++ internal/ingest/ingest.go
@@ -22,11 +22,11 @@ import (
 	indexer "github.com/ipni/go-indexer-core"
 	coremetrics "github.com/ipni/go-indexer-core/metrics"
 	"github.com/ipni/storetheindex/api/v0/ingest/schema"
+	"github.com/ipni/storetheindex/carwriter"
 	"github.com/ipni/storetheindex/config"
 	"github.com/ipni/storetheindex/dagsync"
-	"github.com/ipni/storetheindex/internal/carwriter"
+	"github.com/ipni/storetheindex/filestore"
 	"github.com/ipni/storetheindex/internal/counter"
-	"github.com/ipni/storetheindex/internal/filestore"
 	"github.com/ipni/storetheindex/internal/metrics"
 	"github.com/ipni/storetheindex/internal/registry"
 	"github.com/ipni/storetheindex/peerutil"
